wefixit/backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
gunicorn
python-jose[cryptography]
passlib[bcrypt]
email-validator
bcrypt
dnspython
gunicorn
